server/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
PyMySQL==1.1.0
pymssql==2.2.11
cryptography==41.0.7
PyJWT==2.8.0
python-dotenv==1.0.0
